backend/requirements.txt
fastapi
uvicorn
python-docx
qdrant-client
sentence-transformers
PyPDF2
